Adding held item DURABILITY as a requirement check.
You can now specify a DURABILITY check in a trigger script as a
requirement.  the DURABILITY check requires two parameters.  The first
parameter is a >, <, or = symbol denoting the type of check to perform
against the current item's remaining durability.  The second parameter
is either a simple number value compared against the item's durability,
or a percentage of the item's max durability.  The percentage must be a
3 digit or less number with a percent sign (%) following it, as in this
example:

'Durability Check':
Type: Trigger
Requirements:
Mode: All
List:
- DURABILITY > 50%

The above would trigger if the held item's remaining durability is
greater than 50% of it's max durability.