Scrap "block moves"
@davidcernat - a single daylight detector on my server, with 5 redstone
wires on it, triggers this event roughly 200 times per second. Not quite
sure why... but it's much too spammy to be monitoring this event. Maybe
it could be readded as an event that uh... involves blocks actually
moving, just not a physics recalc.

(Linking you, David, because I assume you added this and will want to be
involved)