add 'determine uncancelled' to 'on player clicks'
A lot of click events seem to be cancelled by default... this adds a
determination for 'uncancelled' in the click event.
Actual usage: right-clicking a dragoncontrolstaff item while riding an
enderdragon:
normally, does nothing due to arbitrarily pre-cancellation (At least on
C-K's server)... but with a 'determine uncancelled' in the event,
right-clicking will properly cause the dragon to move downward.
I imagine plugins will cancel a lot of events that Denizen users want to
not be cancelled... so I think this should be the first of many
'uncancelled' determine options.