Add new meta system, "<@link MetaType MetaName>"
Use this to connect different meta informations together. A meta-parser
should interpret this to the most direct possible link to the
information available in the same location as the meta being read is at.
On websites, it should create a hyperlink with naming text, EG: <a
href="/meta/example">Example Meta</a>    that fits in smoothly with
English phrases like "See [Example Meta] for more information."

Also, let's try to get a lot of these in the meta... Make it all feel
interconnected.