Add Player Jump smartevent
A simple event that was requested to fire when a player jumps into the
air.
Using smart event instead of a normal world event because onPlayerMove
tends to fire very very often, and players jump fairly often... better
to minimize need for this.
There is no proper Bukkit event (thus the player move with weird 'if'
logic)
Cannot really properly block the jump either, so no 'determine
cancelled' is possible for now.
Future possibility ( @aufdemrand ) is 'on player jumps in <notable
cuboid>' or something. Would like to see definite stability in said
system + fully working cuboid-related events before going for it though.