Fix client-crashing PlayEffect bug
This disallows numbers below 1 with iconcrack_[id], which would crash
the Minecraft client.

This commit also changes dInventory#matches to match all arguments
starting with in@, which is the ObjectFetcher prefix. Since every single
dInventory starts with in@, and it always matches in@, that is the only
check necessary.

This commit also fixes some meta.