Always show tags in debug output
This will now show the output indicating how a tag was filled in the
TagManager rather than at the end of Element. Several tags never get
sent to Element and thus wouldn't debug if it weren't for this. An
example of a tag that previously wouldn't debug is any math: tag.
Now, all tags will debug their filling completely regardless of what tag
is used.