Make def[x] return the most-valid object type
Not so sure how to put the meta for that.
The tag will now return the most-valid object type, meaning if the
definition is 'p@player', a dPlayer will be return, if it's 'n@1', an
NPC will be returned, etc... if it's unknown, an Element will be
returned.
Also including a public API function to handle it!
Also,  reorganize element.as_ tags, clean their meta, and add a few
missing ones.