Begin converting doEvents to take dPlayers
dPlayers are needed, most importantly for OfflinePlayers but also just
to fit Denizen standards.
Unfortunately, we can't just have a direct replacement event as dPlayers
and Players are very similar objects in the eyes of Java, and this must
be done in a two-part conversion.
For now, a secondary function using a placeholder event is in use so the
original event can be removed throughout Denizen, Depenizen, and other
API-using projects.
Once that is done in it's entirely ,the placeholder-int function will be
marked deprecated and the old function will be changed to take a dPlayer
instead, and that will be used, until a final commit removes the
placeholder-int function.