First part of UUID support, p@uuid
dPlayer objects now use the UUID to identify, so they can remain valid
even if a player renames.
As a side effect of the renaming system, we can't store player objects
like we used to, and have to go back to gathering player lists from
Bukkit functions and the like.
NEXT UP: Changing the saves.yml file to only use UUIDs...